text_prompt/requirements.txt
absl-py==2.0.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==4.2.2
aniso8601==8.0.0
annotated-types==0.6.0
anyio==3.6.2
asgiref==3.5.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.2.0
backcall==0.2.0
blinker==1.6.2
blis==0.7.11
branca==0.6.0
cachetools==5.3.0
catalogue==2.0.10
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.4
confection==0.1.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.8
dataclasses-json==0.6.4
debugpy==1.6.7.post1
decorator==5.1.1
distlib==0.3.7
distro==1.9.0
dnspython==2.2.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
entrypoints==0.4
etils==1.5.2
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.109.2
filelock==3.12.2
Fiona==1.9.4.post1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==3.0.5
flatbuffers==23.5.26
folium==0.14.0
fonttools==4.39.4
frozenlist==1.4.1
fsspec==2023.10.0
gast==0.4.0
geopandas==0.13.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.23.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.4
imgkit==1.2.3
importlib-metadata==6.4.1
importlib-resources==6.1.1
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jax==0.4.20
jaxlib==0.4.20
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.6
langcodes==3.3.0
langsmith==0.0.87
libclang==16.0.6
lxml==4.9.3
Mako==1.2.4
Markdown==3.5.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
motor==3.1.1
msgpack==1.0.7
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql-connector-python==8.0.28
nest-asyncio==1.5.7
nltk==3.7
numpy==1.25.2
oauthlib==3.2.2
object-detection==0.1
openai==1.12.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
orbax-checkpoint==0.4.4
packaging==23.2
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2022.11.30
platformdirs==3.10.0
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.16.1
pymongo==4.3.3
Pympler==1.0.1
pyparsing==3.0.9
pypiwin32==223
pyproj==3.6.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
regex==2022.6.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.3.4
rsa==4.9
ruamel.yaml.clib==0.2.7
sacrebleu==2.2.0
scikit-learn==1.3.0
scipy==1.11.2
shapely==2.0.1
six==1.16.0
sklearn==0.0.post7
-e git+https://github.com/tensorflow/models@84000f908779829232b7f23ccdfdef41167e460a#egg=slim&subdirectory=research\slim
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.20
srsly==2.4.8
stack-data==0.6.2
starlette==0.36.3
streamlit==1.28.1
tenacity==8.2.3
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.15.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorstore==0.1.45
termcolor==2.3.0
tf-slim==1.1.0
thinc==8.2.3
threadpoolctl==3.2.0
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.64.0
traitlets==5.9.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.14.0
validators==0.20.0
virtualenv==20.24.3
virtualenv-clone==0.5.7
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
weasel==0.3.4
Werkzeug==2.3.7
wget==3.2
wrapt==1.14.1
wurlitzer==3.0.3
yarl==1.9.4
zipp==3.15.0
